CH2/pyproject.toml
[tool.poetry]
name = "ch2"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
python-dotenv = "^1.0.1"
langchain-google-genai = "<3.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
